Добавил абстрактный класс BadWordsAnalyzer, реализующий метод isTriggered для классов наследников
